+++ src/CMakeLists.txt
@@ -1,10 +1,22 @@
-add_library(gotcha SHARED gotcha_utils.c gotcha.c gotcha_auxv.c libc_wrappers.c elf_ops.c hash.c tool.c)
+set(GOTCHA_SOURCES
+  gotcha_utils.c 
+  gotcha.c 
+  gotcha_auxv.c 
+  libc_wrappers.c 
+  elf_ops.c 
+  hash.c 
+  tool.c 
+  library_filters.c
+  gotcha_dl.c
+)
+
+add_library(gotcha SHARED ${GOTCHA_SOURCES})
 set_target_properties(gotcha PROPERTIES SOVERSION ${LIBTOOL_INTERFACE})
 set_target_properties(gotcha PROPERTIES VERSION "${LIBTOOL_INTERFACE}.${LIBTOOL_REVISION}.${LIBTOOL_AGE}")
 set_target_properties(gotcha PROPERTIES COMPILE_FLAGS "-fvisibility=${DEFAULT_SYMBOL_VISIBILITY}")
 set_target_properties(gotcha PROPERTIES LINK_FLAGS "-fvisibility=${DEFAULT_SYMBOL_VISIBILITY}")
 if(GOTCHA_ENABLE_TESTS)
-add_library(gotcha_no_libc SHARED gotcha_utils.c gotcha.c gotcha_auxv.c libc_wrappers.c elf_ops.c hash.c tool.c)
+add_library(gotcha_no_libc SHARED ${GOTCHA_SOURCES})
 set_target_properties(gotcha_no_libc PROPERTIES SOVERSION ${LIBTOOL_INTERFACE})
 set_target_properties(gotcha_no_libc PROPERTIES VERSION "${LIBTOOL_INTERFACE}.${LIBTOOL_REVISION}.${LIBTOOL_AGE}")
 set_target_properties(gotcha_no_libc PROPERTIES COMPILE_DEFINITIONS "FORCE_NO_LIBC")
